+++ src/core/ObservableQuery.ts
@@ -27,7 +27,6 @@ import { NetworkStatus } from '../queries/store';
 
 import { addFragmentsToDocument } from '../queries/getFromAST';
 
-import assign = require('lodash/assign');
 import isEqual = require('lodash/isEqual');
 
 export type ApolloCurrentResult = {
@@ -149,21 +148,26 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
   }
 
   public refetch(variables?: any): Promise<ApolloQueryResult> {
-    this.variables = assign({}, this.variables, variables);
+    this.variables = {
+      ...this.variables,
+      ...variables,
+    };
 
     if (this.options.noFetch) {
       throw new Error('noFetch option should not use query refetch.');
     }
 
     // Update the existing options with new variables
-    assign(this.options, {
-      variables: this.variables,
-    });
+    this.options.variables = {
+      ...this.options.variables,
+      ...this.variables,
+    };
 
     // Override forceFetch for this call only
-    const combinedOptions = assign({}, this.options, {
+    const combinedOptions = {
+      ...this.options,
       forceFetch: true,
-    });
+    };
 
     return this.queryManager.fetchQuery(this.queryId, combinedOptions, FetchType.refetch)
       .then(result => this.queryManager.transformResult(result));
@@ -182,20 +186,26 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
           combinedOptions = fetchMoreOptions;
         } else {
           // fetch the same query with a possibly new variables
-          const variables = assign({}, this.variables, fetchMoreOptions.variables);
+          const variables = {
+            ...this.variables,
+            ...fetchMoreOptions.variables,
+          };
 
-          combinedOptions = assign({}, this.options, fetchMoreOptions, {
+          combinedOptions = {
+            ...this.options,
+            ...fetchMoreOptions,
             variables,
-          });
+          };
         }
 
         // We add the fragments to the document to pass only the document around internally.
         const fullQuery = addFragmentsToDocument(combinedOptions.query, combinedOptions.fragments);
 
-        combinedOptions = assign({}, combinedOptions, {
+        combinedOptions = {
+          ...combinedOptions,
           query: fullQuery,
           forceFetch: true,
-        }) as WatchQueryOptions;
+        } as WatchQueryOptions;
         return this.queryManager.fetchQuery(qid, combinedOptions);
       })
       .then((fetchMoreResult) => {
@@ -263,7 +273,10 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
 
   public setOptions(opts: ModifiableWatchQueryOptions): Promise<ApolloQueryResult> {
     const oldOptions = this.options;
-    this.options = assign({}, this.options, opts) as WatchQueryOptions;
+    this.options = {
+      ...this.options,
+      ...opts,
+    } as WatchQueryOptions;
 
     if (opts.pollInterval) {
       this.startPolling(opts.pollInterval);
@@ -271,8 +284,8 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
       this.stopPolling();
     }
 
-    // If forceFetch went from false to true
-    if (!oldOptions.forceFetch && opts.forceFetch) {
+    // If forceFetch went from false to true or noFetch went from true to false
+    if ((!oldOptions.forceFetch && opts.forceFetch) || (oldOptions.noFetch && !opts.noFetch)) {
       return this.queryManager.fetchQuery(this.queryId, this.options)
         .then(result => this.queryManager.transformResult(result));
     }
@@ -291,16 +304,20 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
    * the previous values of those variables will be used.
    */
   public setVariables(variables: any): Promise<ApolloQueryResult> {
-    const newVariables = assign({}, this.variables, variables);
+    const newVariables = {
+      ...this.variables,
+      ...variables,
+    };
 
     if (isEqual(newVariables, this.variables)) {
       return this.result();
     } else {
       this.variables = newVariables;
       // Use the same options as before, but with new variables
-      return this.queryManager.fetchQuery(this.queryId, assign(this.options, {
+      return this.queryManager.fetchQuery(this.queryId, {
+        ...this.options,
         variables: this.variables,
-      }) as WatchQueryOptions)
+      } as WatchQueryOptions)
         .then(result => this.queryManager.transformResult(result));
     }
   }
@@ -366,6 +383,11 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
 
     const retQuerySubscription = {
       unsubscribe: () => {
+        if (this.observers.findIndex(el => el === observer) < 0 ) {
+          // XXX can't unsubscribe if you've already unsubscribed...
+          // for some reason unsubscribe gets called multiple times by some of the tests
+          return;
+        }
         this.observers = this.observers.filter((obs) => obs !== observer);
 
         if (this.observers.length === 0) {
@@ -435,6 +457,9 @@ export class ObservableQuery extends Observable<ApolloQueryResult> {
     this.subscriptionHandles = [];
 
     this.queryManager.stopQuery(this.queryId);
+    if (this.shouldSubscribe) {
+      this.queryManager.removeObservableQuery(this.queryId);
+    }
     this.observers = [];
   }
 }