+++ test/mutationResults.ts
@@ -5,8 +5,12 @@ import { MutationBehaviorReducerArgs, MutationBehavior, cleanArray } from '../sr
 import { NormalizedCache, StoreObject } from '../src/data/storeUtils';
 import { isMutationResultAction, isQueryResultAction } from '../src/actions';
 
+import { Subscription } from '../src/util/Observable';
+
 import assign = require('lodash/assign');
 import clonedeep = require('lodash/cloneDeep');
+
+import { ObservableQuery } from '../src/core/ObservableQuery';
 
 import gql from 'graphql-tag';
 
@@ -34,6 +38,28 @@ describe('mutation results', () => {
         __typename
         id
         todos {
+          __typename
+          text
+          completed
+        }
+      }
+    }
+  `;
+
+  const queryWithVars = gql`
+    query todoList ($id: Int){
+      __typename
+      todoList(id: $id) {
+        __typename
+        id
+        todos {
+          id
+          __typename
+          text
+          completed
+        }
+        filteredTodos: todos(completed: true) {
+          id
           __typename
           text
           completed
@@ -90,6 +116,68 @@ describe('mutation results', () => {
             completed: false,
           },
         ],
+      },
+    },
+  };
+
+  const result6: any = {
+    data: {
+      __typename: 'Query',
+      todoList: {
+        __typename: 'TodoList',
+        id: '6',
+        todos: [
+          {
+            __typename: 'Todo',
+            id: '13',
+            text: 'Hello world',
+            completed: false,
+          },
+          {
+            __typename: 'Todo',
+            id: '16',
+            text: 'Second task',
+            completed: false,
+          },
+          {
+            __typename: 'Todo',
+            id: '112',
+            text: 'Do other stuff',
+            completed: false,
+          },
+        ],
+        filteredTodos: [],
+      },
+    },
+  };
+
+  const result5: any = {
+    data: {
+      __typename: 'Query',
+      todoList: {
+        __typename: 'TodoList',
+        id: '5',
+        todos: [
+          {
+            __typename: 'Todo',
+            id: '13',
+            text: 'Hello world',
+            completed: false,
+          },
+          {
+            __typename: 'Todo',
+            id: '16',
+            text: 'Second task',
+            completed: false,
+          },
+          {
+            __typename: 'Todo',
+            id: '112',
+            text: 'Do other stuff',
+            completed: false,
+          },
+        ],
+        filteredTodos: [],
       },
     },
   };
@@ -686,6 +774,72 @@ describe('mutation results', () => {
       });
     });
 
+    it('passes variables', () => {
+      let counter = 0;
+      let observableQuery: ObservableQuery;
+      let subscription: any;
+
+      return setup({
+        request: { query: queryWithVars, variables: { id: 5 } },
+        result: result5,
+      }, {
+        request: { query: mutation},
+        result: mutationResult,
+      }, {
+        request: { query: queryWithVars, variables: { id: 6 } },
+        result: result6,
+      }, {
+        request: { query: mutation},
+        result: mutationResult,
+      })
+      .then(() => {
+        observableQuery = client.watchQuery({
+          query: queryWithVars,
+          variables: { id: 5 },
+          reducer: (previousResult, action, variables: any) => {
+            counter++;
+            if (isMutationResultAction(action) && variables['id'] === 5) {
+              const newResult = clonedeep(previousResult) as any;
+              newResult.todoList.todos.unshift(action.result.data.createTodo);
+              return newResult;
+            }
+            return previousResult;
+          },
+        });
+
+        subscription = observableQuery.subscribe({
+          next: () => null, // TODO: we should actually check the new result
+        });
+        return client.mutate({
+          mutation,
+        });
+      })
+      .then(() => {
+        return observableQuery.setOptions({ variables: { id: 6 } });
+      })
+      .then((res) => {
+        return client.mutate({
+          mutation,
+        });
+      })
+      .then(() => {
+        // going back to check the result of the original query
+        return observableQuery.setOptions({ variables: { id: 5 } });
+      })
+      .then((newResult: any) => {
+        subscription.unsubscribe();
+
+        // The reducer should have been called twice
+        assert.equal(counter, 3);
+
+        // But there should be one more todo item than before, because variables only matched once
+        assert.equal(newResult.data.todoList.todos.length, 4);
+
+        // Since we used `prepend` it should be at the front
+        assert.equal(newResult.data.todoList.todos[0].text, 'This one was created with a mutation.');
+      });
+    });
+
     it('can filter based on operationName', () => {
       let counter = 0;
       let observableQuery: any;
@@ -993,9 +1147,19 @@ describe('mutation results', () => {
     };
 
     it('analogous of ARRAY_INSERT', () => {
+      let subscriptionHandle: Subscription;
       return setup({
         request: { query: mutation },
         result: mutationResult,
+      })
+      .then(() => {
+        // we have to actually subscribe to the query to be able to update it
+        return new Promise( (resolve, reject) => {
+          const handle = client.watchQuery({ query });
+          subscriptionHandle = handle.subscribe({
+            next(res) { resolve(res); },
+          });
+        });
       })
       .then(() => {
         return client.mutate({
@@ -1017,6 +1181,8 @@ describe('mutation results', () => {
         return client.query({ query });
       })
       .then((newResult: any) => {
+        subscriptionHandle.unsubscribe();
+
         // There should be one more todo item than before
         assert.equal(newResult.data.todoList.todos.length, 4);
 
@@ -1105,9 +1271,19 @@ describe('mutation results', () => {
         errors.push(msg);
       };
 
+      let subscriptionHandle: Subscription;
       return setup({
         request: { query: mutation },
         result: mutationResult,
+      })
+      .then(() => {
+        // we have to actually subscribe to the query to be able to update it
+        return new Promise( (resolve, reject) => {
+          const handle = client.watchQuery({ query });
+          subscriptionHandle = handle.subscribe({
+            next(res) { resolve(res); },
+          });
+        });
       })
       .then(() => {
         return client.mutate({
@@ -1120,6 +1296,7 @@ describe('mutation results', () => {
         });
       })
       .then(() => {
+        subscriptionHandle.unsubscribe();
         assert.lengthOf(errors, 1);
         assert.equal(errors[0].message, `Hello... It's me.`);
         console.error = oldError;